compare model built from script with one from libProto
